mediacapture-extensions/MediaStreamTrack-getSettings.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/mediacapture-image/resources/imagecapture-helpers.js"></script>
<script>

// This test verifies that the settings defined in the mock Mojo service
// implementation are the same as those returned by the corresponding
// MediaStreamTrack.getSettings().

image_capture_test(async (t, imageCaptureTest) => {
  const mockState = imageCaptureTest.mockImageCapture().state();

  // |stream| must be created _after_ |mock| is constructed to give the
  // latter time to override the bindings.
  const stream = await navigator.mediaDevices.getUserMedia({video: true});
  const [videoTrack] = stream.getVideoTracks();

  // |videoTrack|s settings retrieval, just like the actual capture, is a
  // process kicked right after creation, we introduce a small delay to
  // allow for those to be collected.
  await new Promise(resolve => step_timeout(resolve, 100));

  const settings = videoTrack.getSettings();
  assert_equals(typeof settings, 'object');

  assert_equals(settings.backgroundBlur, mockState.backgroundBlur,
                'backgroundBlur');
}, "exercises MediaStreamTrack.getSettings()");
</script>
